continuity-ai-prototype/AI/requirements.txt
# Core API dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
python-dotenv==1.0.1

# Entity extraction (NER) and fact validation dependencies
# transformers 4.36.2+ supports microsoft/deberta-large-mnli for NLI-based fact validation
transformers==4.36.2
torch==2.1.2

# RAG system dependencies
chromadb==0.5.23
sentence-transformers==3.3.1
llama-cpp-python==0.3.16

# Optional: Discord bot integration
# discord.py==2.4.0
# aiohttp==3.11.11
# asyncio-contextmanager==1.0.1
